refactor: move startSlide function to helper
