+++ src/features/landing/components/ImageSlide.tsx
@@ -1,12 +1,8 @@
 import { FunctionComponent, useEffect, useRef } from 'react';
 
 import usePromises from '~/features/landing/hooks/usePromises';
-import { drawImageFrame, fetchImage } from '~/features/landing/helper';
-import {
-  IMAGE_SLIDE_DELAY,
-  PHOTO_PATH_PREFIX,
-  IMAGE_SLIDE_SPEED,
-} from '~/features/landing/constants';
+import { fetchImage, startSlide } from '~/features/landing/helper';
+import { PHOTO_PATH_PREFIX } from '~/features/landing/constants';
 import { imageSlideElementList } from '~/features/landing/mocks';
 import { resizeCanvasToCoverWindow } from '~/features/shared/utils/canvas';
 
@@ -18,26 +14,6 @@ const ImageSlide: FunctionComponent = () => {
     ),
   );
 
-  const startSlide = () => {
-    const canvas = ref.current;
-    const ctx = canvas?.getContext('2d', { alpha: false });
-    if (!canvas || !ctx || !images) return;
-
-    const canvasSize = { width: canvas.width, height: canvas.height };
-
-    const drawAnimationFrame = (dy: number) => {
-      if (dy > canvas.height) return;
-
-      ctx.clearRect(0, 0, canvas.width, canvas.height);
-      images.forEach((image, idx) =>
-        drawImageFrame(ctx, image, dy, IMAGE_SLIDE_DELAY * idx, canvasSize),
-      );
-      requestAnimationFrame(() => drawAnimationFrame(dy + IMAGE_SLIDE_SPEED));
-    };
-
-    drawAnimationFrame(0);
-  };
-
   useEffect(() => {
     const canvas = ref.current;
     if (canvas === null) return;
@@ -45,7 +21,9 @@ const ImageSlide: FunctionComponent = () => {
   }, []);
 
   useEffect(() => {
-    if (images) startSlide();
+    if (images && ref.current) {
+      startSlide(images, ref?.current);
+    }
   }, [images]);
 
   return (
