+++ src/features/landing/helper.ts
@@ -1,4 +1,8 @@
-import { IMAGE_SLIDE_ACC } from '~/features/landing/constants';
+import {
+  IMAGE_SLIDE_ACC,
+  IMAGE_SLIDE_DELAY,
+  IMAGE_SLIDE_SPEED,
+} from '~/features/landing/constants';
 import { CanvasSize } from '~/features/landing/types';
 
 export const fetchImage = (src: string) => {
@@ -17,7 +21,7 @@ const calcYDistance = (dt: number, a: number): number => {
   return Math.floor((a * dt * dt) / 2);
 };
 
-export const drawImageFrame = (
+const drawImageFrame = (
   ctx: CanvasRenderingContext2D,
   img: HTMLImageElement,
   dt: number,
@@ -40,4 +44,26 @@ export const drawImageFrame = (
     width,
     img.height,
   );
+};
+
+export const startSlide = (
+  images: HTMLImageElement[],
+  canvas: HTMLCanvasElement,
+) => {
+  const ctx = canvas?.getContext('2d', { alpha: false });
+  if (!canvas || !ctx || !images) return;
+
+  const canvasSize = { width: canvas.width, height: canvas.height };
+
+  const drawAnimationFrame = (dy: number) => {
+    if (dy > canvas.height) return;
+
+    ctx.clearRect(0, 0, canvas.width, canvas.height);
+    images.forEach((image, idx) =>
+      drawImageFrame(ctx, image, dy, IMAGE_SLIDE_DELAY * idx, canvasSize),
+    );
+    requestAnimationFrame(() => drawAnimationFrame(dy + IMAGE_SLIDE_SPEED));
+  };
+
+  drawAnimationFrame(0);
 };
